Fix: error foreign key and change DB to mysql

+++ lawchatapp/database/migrations/2024_10_16_103614_create_chats_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('chats', function (Blueprint $table) {
             $table->id();
-            $table->integer('chatroom_id');
+            $table->unsignedBigInteger('chatroom_id');
             $table->foreign('chatroom_id')->references('id')->on('chatrooms');
             $table->timestamps();
         });
